+++ apps/vue/src/App.vue
@@ -9,6 +9,24 @@ const FLATFILE_PUBLISHABLE_KEY = 'your_key'
 export default defineComponent({
   setup() {
     const publishableKey = FLATFILE_PUBLISHABLE_KEY
+    
+    // Define themeConfig
+    const themeConfig = {
+      root: {
+        primaryColor: "#3ccc65",
+        backgroundColor: "#ffffff", 
+        textColor: "#1f2937",
+        borderRadius: "8px",
+      },
+      sidebar: {
+        logo: "https://images.ctfassets.net/hjneo4qi4goj/gL6Blz3kTPdZXWknuIDVx/7bb7c73d93b111ed542d2ed426b42fd5/flatfile.svg",
+        backgroundColor: "#3ccc65",
+        textColor: "#1f2937",
+        width: "280px",
+        padding: "24px",
+      }
+    }
+    
     const spaceProps = ref({
       // apiUrl: 'https://platform.flatfile.com/api',
       name: 'Trste!',
@@ -16,6 +34,7 @@ export default defineComponent({
       workbook,
       listener,
       externalActorId: 'test-1',
+      themeConfig,
       document: {
         title: 'Instructions',
         body:
@@ -24,8 +43,6 @@ export default defineComponent({
           'Please only import CSV and Excel files',
         defaultPage: true,
       },
-
-      themeConfig: { primaryColor: '#546a76', textColor: '#fff' },
       userInfo: {
         name: 'my space name',
       },