+++ packages/cli/src/x/actions/delete.action.ts
@@ -1,10 +1,10 @@
 import chalk from 'chalk'
 import { program } from 'commander'
 import ora from 'ora'
-import { apiKeyClient } from './auth.action'
+import prompts from 'prompts'
 import { getAuth } from '../../shared/get-auth'
 import { messages } from '../../shared/messages'
-import { prompt } from 'prompts'
+import { apiKeyClient } from './auth.action'
 
 export async function deleteAction(
   options?: Partial<{
@@ -61,7 +61,7 @@ export async function deleteAction(
 
     validatingSpinner.succeed(`Found agent match, ${agent.slug}`)
 
-    const { deleteAgent } = await prompt({
+    const { deleteAgent } = await prompts({
       type: 'confirm',
       name: `deleteAgent`,
       message: chalk.red(